recompute support rng
