+++ python/paddle/autograd/py_layer.py
@@ -176,9 +176,8 @@ class CPyLayer(object):
 
 class PyLayerBackward(PyLayerContext):
     def backward(self, *args, **kwargs):
-        # with paddle.fluid.dygraph.no_grad():
-        # with paddle.fluid.dygraph.guard():
-        return self._forward_cls.backward(*args, **kwargs)
+        with paddle.fluid.dygraph.no_grad():
+            return self._forward_cls.backward(*args, **kwargs)
 
 
 class LayerMeta(type):
