+++ python/paddle/utils/__init__.py
@@ -21,7 +21,7 @@ from .op_version import OpLastCheckpointChecker
 from .install_check import run_check
 from ..fluid.framework import unique_name
 from ..fluid.framework import require_version
-from .checkpoint import checkpoint
+from .recompute import recompute
 
 from . import download
 
@@ -32,4 +32,4 @@ __all__ = ['dump_config', 'deprecated', 'download', 'run_check']
 #TODO: define new api under this directory
 __all__ += ['unique_name', 'require_version']
 
-__all__ += ['checkpoint']
+__all__ += ['recompute']
